Safari-611.1.10/LayoutTests/fast/shadow-dom/fullscreen-in-shadow-event-should-propagate-expected.txt

Test that webkitfullscreenchange from shadow tree should be propagated to the shadow host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS () => event.composed is true
PASS () => event.currentTarget is [object ShadowRoot]
PASS () => event.composed is true
PASS () => event.currentTarget is [object HTMLDocument]
PASS successfullyParsed is true

TEST COMPLETE
Enter fullscreen
